+++ tests/test_app.rs
@@ -93,10 +93,7 @@ fn test_simple_app() {
                           "format": "date-time",
                           "type": "string"
                         },
-                        "uid": {
-                          "format": "uuid",
-                          "type": "string"
-                        }
+                        "uid": {}
                       },
                       "required":["class", "name", "updated"]
                     }
@@ -537,10 +534,7 @@ fn test_list_in_out() {
                           "format": "date-time",
                           "type": "string"
                         },
-                        "uid": {
-                          "format": "uuid",
-                          "type": "string"
-                        }
+                        "uid": {}
                       },
                       "required":["class", "name", "updated"]
                     }
@@ -642,10 +636,7 @@ fn test_impl_traits() {
                           "format": "date-time",
                           "type": "string"
                         },
-                        "uid": {
-                          "format": "uuid",
-                          "type": "string"
-                        }
+                        "uid": {}
                       },
                       "required":["class", "name", "updated"]
                     }